Predictive Data Analysis based on K Neighbors
This model uses one of the most efficient libraries of python i.e. scikit. The data imported is a multivariate dataset introduced by the British statistician and biologist Ronald Fisher in his 1936 paper.  The dataset has four types of features viz. Petal length, Petal width, Sepal Length, Sepal width,  and one kind  of label which determines whether the flower is Setosa, Versicolor, and virginica. In order to train and test the dataset. The same dataset is being used by diving it in a ratio of 0.5. The model uses K Neighbor analysis. The accuracy achieved while test is 97%.